requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pika==1.3.2
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
asyncpg==0.29.0
httpx==0.25.2
requests==2.31.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
nanoid==2.0.0
